requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
environs==10.3.0
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
isort==5.13.2
magic-filter==1.0.12
marshmallow==3.20.2
mccabe==0.7.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
